Fix safeStringify redaction (oh-tab-0sac)
